Типы перенесены в папку shared
