+++ markup/favorites-empty.html
