+++ src/cli/commands/import.command.ts
@@ -1,6 +1,6 @@
 import { exit } from 'node:process';
 import { TSVFileReader } from '../../shared/libs/file-reader/index.js';
-import { TUser } from '../../types/index.js';
+import { TUser } from '../../shared/types/index.js';
 import { ICommand } from './command.interface.js';
 import chalk from 'chalk';
