+++ src/shared/libs/file-reader/file-reader.interface.ts
