+++ src/shared/libs/file-reader/index.ts
