+++ src/shared/types/index.ts
@@ -1,0 +1,4 @@
+export { TComment } from './comment.type.js';
+export { TUser } from './user.type.js';
+export { TOfferType, TOffer } from './offer.type.js';
+export { TCity, TCityName } from './city.type.js';
