+++ src/shared/types/user.type.ts
@@ -1,0 +1,9 @@
+type TUser = {
+  name: string;
+  email: string;
+  avatarUrl?: string;
+  password: string;
+  isPro: boolean;
+};
+
+export {TUser};
