--- src/types/comment.type.ts
@@ -1,11 +1,0 @@
-import { TUser } from './user.type.js';
-
-type TComment = {
-  id: string;
-  offerId: string;
-  comment: string;
-  date: string;
-  user: TUser;
-};
-
-export type { TComment };
