--- src/types/user.type.ts
@@ -1,9 +1,0 @@
-type TUser = {
-  name: string;
-  email: string;
-  avatarUrl?: string;
-  password: string;
-  isPro: boolean;
-};
-
-export {TUser};
